cmd/compile: don't treat an InlMark as a read during deadstore
An InlMark "read" can't make an otherwise dead store live. Without this
CL, we sometimes zero an object twice in succession because we think
there is a reader in between.

Kind of challenging to make a test for this. The second zeroing has the
same instruction on the same line number, so codegen tests can't see it.

Fixes #67957

+++ src/runtime/symtabinl_test.go
@@ -106,8 +106,8 @@ func lineNumber() int {
 
 // Below here is the test data for XTestInlineUnwinder
 
-var tiuStart = lineNumber() // +0
-var tiu1, tiu2, tiu3 int    // +1
+var tiuStart = lineNumber()  // +0
+var tiu1, tiu2, tiu3 float64 // +1. Note float64 to avoid combining two ++ into a +=2.
 func tiuInlined1() { // +2
 	tiu1++ // +3
 } // +4
